refactor: 이메일 전송 기능 역할 분리
